requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
Brotli==1.0.9
CacheControl==0.12.6
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cloudinary==1.25.0
cryptography==3.4.7
dj-database-url==0.5.0
Django==3.2.5
django-appconf==1.0.4
django-bmemcached==0.2.4
django-cachalot==2.4.2
django-cloudinary-storage==0.3.0
django-compressor==2.4.1
django-cors-headers==3.8.0
django-crispy-forms==1.11.2
django-heroku==0.3.1
django-htmlmin==0.11.0
django-pwa==1.0.10
django-recaptcha==2.0.6
django-utils-six==2.0
firebase-admin==5.0.0
gcloud==0.18.3
google-api-core==1.26.3
google-api-python-client==2.1.0
google-auth==1.28.0
google-auth-httplib2==0.1.0
google-cloud-core==1.6.0
google-cloud-firestore==2.1.0
google-cloud-storage==1.40.0
google-crc32c==1.1.2
google-resumable-media==1.3.1
googleapis-common-protos==1.53.0
grpcio==1.36.1
gunicorn==20.1.0
html5lib==1.1
httplib2==0.19.1
idna==2.10
jwcrypto==0.8
jws==0.1.3
msgpack==1.0.2
oauth2client==4.1.3
packaging==20.9
Pillow==8.2.0
proto-plus==1.18.1
protobuf==3.15.6
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
pylibmc==1.5.1
pyparsing==2.4.7
Pyrebase4==4.4.3
python-binary-memcached==0.30.1
python-decouple==3.4
python-jwt==2.0.1
pytz==2021.1
rcssmin==1.0.6
requests==2.25.1
requests-toolbelt==0.9.1
rjsmin==1.1.0
rsa==4.7.2
selenium==3.141.0
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
uhashring==2.1
uritemplate==3.0.1
urllib3==1.26.5
webencodings==0.5.1
whitenoise==5.3.0
